backend/requirements.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
motor>=3.3.0
pymongo>=4.6.0

# ML & Image Processing
tensorflow>=2.12.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# PDF Generation
reportlab>=4.0.0

# Utils
python-dotenv>=1.0.0
aiofiles>=23.0.0
